move Ref, RefMut, and LoanError into js module
