+++ src/js/binary.rs
@@ -3,10 +3,10 @@ use std::mem;
 use std::os::raw::c_void;
 use std::slice;
 use vm::{Context, VmResult};
-use js::{Value, Object, Borrow, BorrowMut, build};
+use js::{Value, Object, Borrow, BorrowMut, Ref, RefMut, LoanError, build};
 use js::internal::ValueInternal;
 use mem::{Handle, Managed};
-use vm::{VmGuard, Ref, RefMut, LoanError};
+use vm::VmGuard;
 use vm::internal::Pointer;
 use neon_runtime;
 use neon_runtime::raw;
