+++ src/js/class.rs
@@ -5,9 +5,9 @@ use std::os::raw::c_void;
 use neon_runtime;
 use neon_runtime::raw;
 use mem::{Handle, Managed};
-use vm::{JsResult, VmResult, Throw, This, Context, VmGuard, Ref, RefMut, LoanError, Callback};
+use vm::{JsResult, VmResult, Throw, This, Context, VmGuard, Callback};
 use vm::internal::Isolate;
-use js::{Value, JsFunction, Object, JsValue, Borrow, BorrowMut, build};
+use js::{Value, JsFunction, Object, JsValue, Borrow, BorrowMut, Ref, RefMut, LoanError, build};
 use js::internal::ValueInternal;
 use js::error::{JsError, Kind};
 use self::internal::{ClassMetadata, MethodCallback, ConstructorCallCallback, AllocateCallback, ConstructCallback};
