+++ src/js/mod.rs
@@ -4,14 +4,16 @@ pub mod binary;
 pub mod error;
 pub mod class;
 
+use std::fmt;
 use std::mem;
 use std::os::raw::c_void;
 use std::marker::PhantomData;
+use std::ops::{Deref, DerefMut, Drop};
 use neon_runtime;
 use neon_runtime::raw;
 use neon_runtime::tag::Tag;
 use mem::{Handle, Managed};
-use vm::{Context, VmGuard, Ref, RefMut, LoanError, CallContext, Callback, VmResult, Throw, JsResult, This};
+use vm::{Context, VmGuard, CallContext, Callback, VmResult, Throw, JsResult, This};
 use vm::internal::{Isolate, Pointer};
 use js::error::{JsError, Kind};
 use self::internal::{ValueInternal, SuperType, FunctionCallback};
@@ -745,4 +747,88 @@ pub trait BorrowMut: Borrow {
     }
 
     fn try_borrow_mut<'a>(self, guard: &'a VmGuard<'a>) -> Result<RefMut<'a, Self::Target>, LoanError>;
+}
+
+pub enum LoanError {
+    Mutating(*const c_void),
+    Frozen(*const c_void)
+}
+
+impl fmt::Display for LoanError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match *self {
+            LoanError::Mutating(p) => {
+                write!(f, "outstanding mutable loan exists for object at {:?}", p)
+            }
+            LoanError::Frozen(p) => {
+                write!(f, "object at {:?} is frozen", p)
+            }
+        }
+    }
+}
+
+// FIXME: this should be covariant in 'a and T -- is it?
+// https://doc.rust-lang.org/nomicon/subtyping.html
+pub struct Ref<'a, T: Pointer> {
+    pointer: T,
+    guard: &'a VmGuard<'a>
+}
+
+impl<'a, T: Pointer> Ref<'a, T> {
+    pub(crate) unsafe fn new(guard: &'a VmGuard<'a>, pointer: T) -> Result<Self, LoanError> {
+        let mut ledger = guard.ledger.borrow_mut();
+        ledger.try_borrow(pointer.as_ptr())?;
+        Ok(Ref { pointer, guard })
+    }
+}
+
+impl<'a, T: Pointer> Drop for Ref<'a, T> {
+    fn drop(&mut self) {
+        let mut ledger = self.guard.ledger.borrow_mut();
+        ledger.settle(unsafe { self.pointer.as_ptr() });
+    }
+}
+
+impl<'a, T: Pointer> Deref for Ref<'a, T> {
+    type Target = T;
+
+    fn deref(&self) -> &Self::Target {
+        &self.pointer
+    }
+}
+
+// FIXME: I think this should be invariant in T -- is it?
+// https://doc.rust-lang.org/nomicon/subtyping.html
+pub struct RefMut<'a, T: Pointer> {
+    pointer: T,
+    guard: &'a VmGuard<'a>
+}
+
+impl<'a, T: Pointer> RefMut<'a, T> {
+    pub(crate) unsafe fn new(guard: &'a VmGuard<'a>, mut pointer: T) -> Result<Self, LoanError> {
+        let mut ledger = guard.ledger.borrow_mut();
+        ledger.try_borrow_mut(pointer.as_mut())?;
+        Ok(RefMut { pointer, guard })
+    }
+}
+
+impl<'a, T: Pointer> Drop for RefMut<'a, T> {
+    fn drop(&mut self) {
+        let mut ledger = self.guard.ledger.borrow_mut();
+        ledger.settle_mut(unsafe { self.pointer.as_mut() });
+    }
+}
+
+impl<'a, T: Pointer> Deref for RefMut<'a, T> {
+    type Target = T;
+
+    fn deref(&self) -> &Self::Target {
+        &self.pointer
+    }
+}
+
+impl<'a, T: Pointer> DerefMut for RefMut<'a, T> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.pointer
+    }
 }
